Fix for typo made during compatability upgrades
Add Django-object-tools and django-export for admin ui goodness, allows easy dumping of data to multiple formats.
NOTE DJANGO-OBJECT-TOOLS AS IS IS NOT COMPATIBLE WITH 1.8. Using an alternative version with the necessary fixes on github until the PR fixing it is merged.

Use flat-theme for the admin ui

Add commented out django-redist functionality for caching